requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
beautifulsoup4
requests